Include Content Server object ids in doc ids.
The adaptor was using the object path as the doc id and then using the
path to look the object up when retrieving content. Path lookup
fails for objects in volumes (such as Projects). Include the object id
in the doc id for correct lookup and add a separate step to verify
that the doc id/path still matches the object's actual path.