Support all node types.
For any node type, send the metadata and at least a dummy HTML
document with the node name.

For types with custom content objects (mostly in the
Collaboration service), use those objects to send the
type-specific content. For types that are also containers, if the
attempt to retrieve the custom content object with the
Collaboration service fails, send the children as if the object
were a generic container.

References are not dereferenced: the content of an Alias or URL
is not indexed. Collections are treated as lists of names, but
the contents are not treated as document or folder content to be
indexed. The document version referenced by a Generation node is
not indexed.

http://codereview.appspot.com/239120043