Support configuration for default version number.
In getVersion, there are three special version numbers: 0 for the most
recent version overall (major or minor), -1 for the most recent major
version of a node with major/minor versioning, and -2 for the most
recent version (major only for major/minor versioning nodes). Extra
permissions are required to see minor versions, so use -2 as the
default.

https://codereview.appspot.com/243530043/