Add option to skip indexing folders.
Add opentext.indexFolders configuration parameter. When this is set to
false, folders will be traversed but not indexed. If the Folder type
is added to excludedNodeTypes, it will act as if indexFolders had been
set to false.

Moved the setting of displayUrl to getDocContent. This was done first
to simplify the process of not setting it for folders that aren't
being indexed. Even though we actually want it set then, it makes
sense to leave it here now.

http://codereview.appspot.com/315050043