Add opentext.pushLocalGroupsOnly (b/64085002) (#20)
* Add opentext.pushLocalGroupsOnly (b/64085002)

Add option to push only local groups from Content Server to the
GSA. This would apply when other groups are defined elsewhere (LDAP,
etc) and synced to the GSA outside this connector. The Content Web
Services API doesn't give us any information about the source of a
group, so this implementation is limited to checking for a group
that's defined with a Windows domain as part of the name.